data-retrieval/requirements.txt
requests
beautifulsoup4
selenium
python-dotenv
langchain-huggingface
numpy